reafactor: simplify for first version
