+++ lua/obsidian/actions.lua
@@ -622,6 +622,42 @@ M.add_property = function()
   note:update_frontmatter(0)
 end
 
+---@param template_name string
+M.insert_template = function(template_name)
+  local templates_dir = api.templates_dir()
+  if not templates_dir then
+    return log.err "Templates folder is not defined or does not exist"
+  end
+  local templates = require "obsidian.templates"
+
+  -- We need to get this upfront before the picker hijacks the current window.
+  local insert_location = api.get_active_window_cursor_location()
+
+  local function insert_template(name)
+    templates.insert_template {
+      type = "insert_template",
+      template_name = name,
+      template_opts = Obsidian.opts.templates,
+      templates_dir = templates_dir,
+      location = insert_location,
+    }
+  end
+
+  if template_name then
+    insert_template(template_name)
+    return
+  end
+
+  Obsidian.picker.find_files {
+    prompt_title = "Templates",
+    dir = templates_dir,
+    no_default_mappings = true,
+    callback = function(path)
+      insert_template(path)
+    end,
+  }
+end
+
 M.start_presentation = function(buf)
   local note = Note.from_buffer(buf)
   require("obsidian.slides").start_presentation(note)
