+++ lua/obsidian/commands/new_from_template.lua
@@ -1,8 +1,8 @@
-local api = require "obsidian.api"
+local actions = require "obsidian.actions"
 
 ---@param data obsidian.CommandArgs
 return function(data)
   local id = table.concat(data.fargs, " ", 1, #data.fargs - 1)
   local template = data.fargs[#data.fargs]
-  api.new_from_template(id, template)
+  actions.new_from_template(id, template)
 end
