+++ lua/obsidian/commands/template.lua
@@ -1,44 +1,10 @@
-local templates = require "obsidian.templates"
-local log = require "obsidian.log"
-local api = require "obsidian.api"
+local actions = require "obsidian.actions"
 
 ---@param data obsidian.CommandArgs
 return function(data)
-  local templates_dir = api.templates_dir()
-  if not templates_dir then
-    return log.err "Templates folder is not defined or does not exist"
-  end
-
-  -- We need to get this upfront before the picker hijacks the current window.
-  local insert_location = api.get_active_window_cursor_location()
-
-  local function insert_template(name)
-    templates.insert_template {
-      type = "insert_template",
-      template_name = name,
-      templates_dir = templates_dir,
-      location = insert_location,
-    }
-  end
-
+  local template_name
   if string.len(data.args) > 0 then
-    local template_name = vim.trim(data.args)
-    insert_template(template_name)
-    return
+    template_name = vim.trim(data.args)
   end
-
-  local picker = Obsidian.picker
-  if not picker then
-    log.err "No picker configured"
-    return
-  end
-
-  picker.find_files {
-    prompt_title = "Templates",
-    dir = templates_dir,
-    no_default_mappings = true,
-    callback = function(path)
-      insert_template(path)
-    end,
-  }
+  actions.insert_template(template_name)
 end
