+++ lua/obsidian/lsp/handlers/code_action.lua
@@ -1,16 +1,16 @@
 local actions = require("obsidian.lsp.handlers._code_action").actions
 
----@param acts lsp.CodeAction[]
+---@param code_actions lsp.CodeAction[]
 ---@param in_selection boolean
 ---@return string[]
-local function get_commands_by_context(acts, in_selection)
+local function get_commands_by_context(code_actions, in_selection)
   return vim
-    .iter(acts)
-    :filter(function(act)
+    .iter(code_actions)
+    :filter(function(code_action)
       if in_selection then
-        return act.data.range ~= nil
+        return code_action.data.range ~= nil
       else
-        return act.data.range == nil
+        return code_action.data.range == nil
       end
     end)
     :totable()
@@ -19,16 +19,8 @@ end
 ---@param params lsp.CodeActionParams
 return function(params, handler)
   local range = params.range
-
   local in_selection = range.start ~= range["end"]
-
   local res = get_commands_by_context(actions, in_selection)
-
-  vim.tbl_map(function(act)
-    if act.data.edit and in_selection then
-      act.edit = act.data.edit(range)
-    end
-  end, res)
 
   handler(nil, res)
 end
